+++ docs/components/guide/GuideFieldSelects.vue
@@ -42,7 +42,7 @@
       code(class="html")
         | &lt;!-- Insert this component in your code --&gt;
         | &lt;!-- Customize it with props (see table below) --&gt;
-        | &lt;dm-select&gt;
+        | &lt;dm-select v-model="synchronizedValue"&gt;
         |   &lt;template #option-left="{ option }"&gt;&lt;/template&gt;
         |   &lt;template #option-right="{ option }"&gt;&lt;/template&gt;
         | &lt;/dm-select&gt;